refactor(config): move the config functionality for the commons package
This commit moves the config functionality from the server package to
the commons package, as the seeding and rate updating were moved into
separate services it wasn't only the server using the config anymore and
it made sense to move it to the commons package. The changes are just
changes of which package is being called
