+++ internal/commons/config.go
@@ -1,4 +1,4 @@
-package server
+package commons
 
 import (
 	"fmt"
